pyproject.toml
[tool.poetry]
name = "olympiad-platform"
version = "0.1.0"
description = ""
authors = ["Kseniia Kornilova <ksyu.kornilova2809@gmail.com>"]
readme = "README.md"
packages = [{include = "olympiad_platform"}]

[tool.poetry.dependencies]
python = "^3.12"
django = "^4.2.3"
psycopg2 = "^2.9.6"
bootstrap4 = "^0.1.0"
django-bootstrap4 = "^23.2"
django-oauth-toolkit = "^2.3.0"
django-cors-headers = "^4.2.0"
markdown = "^3.6"
social-auth-app-django = "^5.4.2"
pymemcache = "^4.0.0"
django-debug-toolbar = "^4.4.6"
python-dotenv = "^1.0.1"
pillow = "^10.4.0"
djangorestframework = "^3.15.2"
drf-yasg = "^1.21.7"
setuptools = "^73.0.1"
pytest = "^8.3.2"
pytest-django = "^4.8.0"
flake8-import-order = "^0.18.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
